+++ restaurant - Copy/restaurant/classes/Products.cs
@@ -10,7 +10,7 @@ namespace restaurant
     {
         public int id { get; set; }
         public string name { get; set; }
-        public int price { get; set; }
+        public float price { get; set; }
         public string description { get; set; }
         public int category { get; set; }
